BlochSpins.jl

Models

BlochSpins.init_figure — Function

init_figure()

Initialize a 3D figure for Bloch sphere visualization. it returns a tuple (fig, ax) where fig is the Figure and ax is the LScene axis.

BlochSpins.visualize_axis! — Function

visualize_axis!(fig, ax)

Draws the X, Y, Z axes in the given axis ax of the figure fig.

BlochSpins.visualize_bloch_sphere! — Function

visualize_bloch_sphere!(fig, ax)

Draws a Bloch sphere in the given axis ax of the figure fig.

BlochSpins.visualize_vector! — Function

visualize_vector!(fig, ax, vec)

Draws a vector vec originating from the origin in the given axis ax of the figure fig.